requirements.txt
python-dotenv==0.7.1
requests==2.13.0